+++ src/revert/table_account_preference_event_size.sql
@@ -1,0 +1,7 @@
+-- Revert maevsi:table_maevsi.account_preference_event_size from pg
+
+BEGIN;
+
+DROP TABLE maevsi.account_preference_event_size;
+
+COMMIT;